Release Sprint3 avance

+++ app/src/main/java/co/edu/uniandes/misw4203/equipo11/vinilos/data/database/daos/CollectorDAO.kt
@@ -7,11 +7,16 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Transaction
+import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.models.Album
 import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.models.Collector
+import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.models.CollectorAlbum
+import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.models.CollectorAlbumCrossRef
 import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.models.CollectorFavoritePerformer
 import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.models.CollectorWithPerformers
 import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.models.Performer
+import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.toAlbum
 import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.toCollector
+import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.toCollectorAlbum
 import co.edu.uniandes.misw4203.equipo11.vinilos.data.database.toPerformer
 import co.edu.uniandes.misw4203.equipo11.vinilos.data.network.models.CollectorJson
 import kotlinx.coroutines.flow.Flow
@@ -19,9 +24,16 @@ import kotlinx.coroutines.flow.map
 import java.text.Normalizer
 
 @Dao
-interface CollectorDAO {
-    @Query("SELECT * FROM collector")
-    fun getCollectors(): Flow<List<Collector>>
+abstract class CollectorDAO {
+    @Query("SELECT * FROM collector WHERE id = :collectorId")
+    abstract fun getCollectorById(collectorId: Int): Flow<Collector?>
+    @Query("SELECT p.* FROM CollectorFavoritePerformer cp JOIN Performer p on cp.performerId = p.id WHERE cp.collectorId == :collectorId ORDER BY p.name COLLATE UNICODE")
+    abstract fun getFavoritePerformers(collectorId: Int): Flow<List<Performer>>
+
+    @Query("SELECT ca.collectorId, ca.price, ca.status, a.id as album_id, a.name as album_name, a.cover as album_cover, " +
+            "a.releaseDate as album_releaseDate, a.description as album_description, " +
+            "a.genre as album_genre, a.recordLabel as album_recordLabel FROM CollectorAlbumCrossRef ca JOIN Album a on ca.albumId = a.id WHERE ca.collectorId == :collectorId ORDER BY a.name COLLATE UNICODE")
+    abstract fun getAlbums(collectorId: Int): Flow<List<CollectorAlbum>>
 
     //
     // Query list of collectors together with their favorite performers
@@ -45,12 +57,11 @@ interface CollectorDAO {
     //   - C3, null
 
     // Internal only helper data class
-    data class CollectorWithPerformer(
+    protected data class CollectorWithPerformer(
         @Embedded(prefix = "collector_") val collector: Collector,
         @Embedded(prefix = "performer_") val performer: Performer?
     )
 
-    // Internal use only
     @Transaction
     @Query(
         "SELECT c.id as collector_id, c.name as collector_name, c.telephone as collector_telephone, c.email as collector_email, " +
@@ -58,7 +69,7 @@ interface CollectorDAO {
         "FROM Collector c LEFT JOIN CollectorFavoritePerformer cp ON cp.collectorId = c.id LEFT JOIN " +
         "Performer p ON cp.performerId = p.id"
     )
-    fun getCollectorsWithPerformer(): Flow<List<CollectorWithPerformer>>
+    protected abstract fun getCollectorsWithPerformer(): Flow<List<CollectorWithPerformer>>
 
     //
     // Query list of collectors together with their favorite performers
@@ -74,7 +85,7 @@ interface CollectorDAO {
     //   - C1, [P1, P2]
     //   - C2, [P2]
     //   - C3, []
-    fun getCollectorsWithPerformers(): Flow<List<CollectorWithPerformers>> {
+    open fun getCollectorsWithPerformers(): Flow<List<CollectorWithPerformers>> {
         return getCollectorsWithPerformer().map { collectors ->
             collectors
                 .groupBy({ it.collector }, { it.performer })
@@ -90,37 +101,52 @@ interface CollectorDAO {
         }
     }
 
-    // Internal use only
-    @Insert
-    suspend fun insertCollectors(collectors: List<Collector>)
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    protected abstract suspend fun insertCollectors(collectors: List<Collector>)
 
-    // Internal use only
     @Query("DELETE FROM collector")
-    suspend fun deleteCollectors()
+    protected abstract suspend fun deleteCollectors()
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertCollectorFavoritePerformers(collectorFavoritePerformers: List<CollectorFavoritePerformer>)
+    abstract suspend fun insertCollectorFavoritePerformers(collectorFavoritePerformers: List<CollectorFavoritePerformer>)
 
-    // Internal use only
     @Query("DELETE FROM collectorfavoriteperformer")
-    suspend fun deleteAllCollectorFavoritePerformer()
+    protected abstract suspend fun deleteAllCollectorFavoritePerformer()
 
     @Delete
-    suspend fun deleteCollectorFavoritePerformer(collectorFavoritePerformer: CollectorFavoritePerformer)
+    abstract suspend fun deleteCollectorFavoritePerformer(collectorFavoritePerformer: CollectorFavoritePerformer)
 
-    // Internal use only
-    // This is necessary, so that the list of collector with their favorite performers can
-    // be updated in a transaction guaranteeing the consistency of the database
+    @Query("DELETE FROM CollectorFavoritePerformer WHERE collectorId = :collectorId")
+    protected abstract suspend fun deleteCollectorFavoritePerformerByCollectorId(collectorId: Int)
+
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertPerformers(performers: List<Performer>)
+    protected abstract suspend fun insertAlbums(albums: List<Album>)
 
-    // Refresh the list of collectors and their favorite performers
-    // To make sure that the database is consistent it is necessary to update the performers
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    abstract suspend fun insertCollectorAlbums(collectorAlbums: List<CollectorAlbumCrossRef>)
+
+    @Query("DELETE FROM CollectorAlbumCrossRef")
+    protected abstract suspend fun deleteCollectorAlbums()
+
+    @Query("DELETE FROM CollectorAlbumCrossRef WHERE collectorId = :collectorId")
+    protected abstract suspend fun deleteCollectorAlbumsByCollectorId(collectorId: Int)
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    protected abstract suspend fun insertPerformers(performers: List<Performer>)
+
+    // Refresh the list of collectors
+    // To make sure that the database is consistent it is necessary to update the favorite artists and albums
     // associated with the collectors as well
+    //
+    // It is necessary to set deleteAll to remove collectors from the local database that are no longer available upstream
     @Transaction
-    suspend fun deleteAndInsertCollectors(collectors: List<CollectorJson>) {
+    open suspend fun deleteAndInsertCollectors(collectors: List<CollectorJson>, deleteAll: Boolean = true) {
+        val albums: MutableList<Album> = mutableListOf()
+        val collectorAlbums: MutableList<CollectorAlbumCrossRef> = mutableListOf()
+
         val performers: MutableList<Performer> = mutableListOf()
         val collectorFavoritePerformers: MutableList<CollectorFavoritePerformer> = mutableListOf()
+
         val mappedCollectors = collectors.map { collector ->
             val favoritePerformers: List<Performer> = requireNotNull(collector.favoritePerformers).map { it.toPerformer() }
             performers.addAll(favoritePerformers)
@@ -130,13 +156,38 @@ interface CollectorDAO {
                 )
             }
 
+            val collectorAlbumsList = requireNotNull(collector.collectorAlbums).map {
+                albums.add(it.album.toAlbum())
+
+                it.toCollectorAlbum(collector.id)
+            }
+            collectorAlbums.addAll(collectorAlbumsList)
+
             collector.toCollector()
         }
 
-        deleteCollectors()
+        if (deleteAll)
+            deleteCollectors()
         insertCollectors(mappedCollectors)
-        deleteAllCollectorFavoritePerformer()
-        insertCollectorFavoritePerformers(collectorFavoritePerformers)
+
         insertPerformers(performers)
+        if (deleteAll)
+            deleteAllCollectorFavoritePerformer()
+        else {
+            mappedCollectors.forEach { collector ->
+                deleteCollectorFavoritePerformerByCollectorId(collector.id)
+            }
+        }
+        insertCollectorFavoritePerformers(collectorFavoritePerformers)
+
+        insertAlbums(albums)
+        if (deleteAll)
+            deleteCollectorAlbums()
+        else {
+            mappedCollectors.forEach { collector ->
+                deleteCollectorAlbumsByCollectorId(collector.id)
+            }
+        }
+        insertCollectorAlbums(collectorAlbums)
     }
 }
